Add test to report installed PPA packages
For SRUs to Ubuntu and the Cloud Archive, bugs need to be updated
with testing results and package versions tested.

Some times these packages are tested in a PPA, as opposed to the
official -proposed pocket.

This will enable getting the package versions from charm units in
an automated way.

Signed-off-by: Frode Nordahl <fnordahl@ubuntu.com>